Issue #117, working context switch with all registers from the button

+++ board/AVR/drivers/Timer/8-bit_TimerCounter0.h
@@ -32,9 +32,8 @@ void timer0Init1ms();
 
 uint8_t timer0Read();
 
-void timer0Delay1ms(uint16_t millis, FnPtr timeoutFn);
-void timer0Start (uint16_t millis, FnPtr timeoutFn);
-
+void timer0Delay1ms(uint32_t millis, FnPtr timeoutFn);
+void timer0Start (uint32_t millis, FnPtr timeoutFn);
 
 uint8_t timer0ReadOVFFlag();
